feat: expand home assistant telemetry

+++ lib/DeviceState/src/DeviceState.h
@@ -234,7 +234,7 @@ struct DeviceState {
 
     unsigned long uptimeMs = 0;
     uint32_t uptimeSeconds = 0;
-    uint16_t uptimeHours = 0;
+    uint32_t uptimeHours = 0;
     uint8_t uptimeMinutes = 0;
     uint8_t uptimeSecondPart = 0;
     char uptimeText[16] = "00:00:00";
